+++ apps/web/next.config.mjs
