+++ apps/web/package.json
@@ -23,7 +23,6 @@
     "@types/react": "18.3.0",
     "@types/react-dom": "18.3.1",
     "@vanilla-extract/next-plugin": "^2.4.8",
-    "eslint": "^8",
     "eslint-config-next": "14.2.21",
     "typescript": "5.5.4"
   }
