+++ package.json
@@ -35,6 +35,7 @@
     "node": ">=18"
   },
   "dependencies": {
-    "@vanilla-extract/css": "^1.17.0"
+    "@vanilla-extract/css": "^1.17.0",
+    "@vanilla-extract/recipes": "^0.5.5"
   }
 }
