+++ packages/ui/package.json
@@ -2,11 +2,15 @@
   "name": "@repo/ui",
   "version": "0.0.0",
   "private": true,
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js"
+    },
+    "./styles": "./dist/index.css"
+  },
   "scripts": {
-    "build": "tsc --emitDeclarationOnly && esbuild src/index.ts --bundle --platform=node --format=esm --outfile=dist/index.mjs && esbuild src/index.ts --bundle --platform=node --format=cjs --outfile=dist/index.js",
+    "build": "tsc && node esbuild.config.mjs",
     "dev": "esbuild src/index.ts --bundle --platform=node --format=esm --outfile=dist/index.mjs --watch",
     "lint": "eslint . --max-warnings 0"
   },
@@ -19,7 +23,8 @@
   "devDependencies": {
     "@types/react": "^18.2.0",
     "@vanilla-extract/esbuild-plugin": "^2.3.5",
-    "esbuild": "^0.24.0"
+    "esbuild": "^0.21.0",
+    "esbuild-plugin-preserve-directives": "^0.0.11"
   },
   "peerDependencies": {
     "react": "^18.2.0"
